Added freezing functionality to the SoWindowHandling. Connect a trigger to the "freezevideo" field to tell the video component to ignore all video flow, notice that this does not affect openvideo. Connect a trigger to the "freezetracking" field to tell the event system to ignore all event flow, notice that this does not affect opentracker.
To unfreeze the tracking system it was necessary to introduce a wildcard key value pair, that is safe from being ignored by the "freezing" (otherwise you wouldnt be able to unfreeze it).

This new functionality allows to take "pictures" of scenes, on video, on tracking or on both. This separation allows the ability of "manual" registration of objects if the tracking is too bad.